requirements.txt
pandas
numpy
matplotlib
Pillow
seaborn
litdata 
tifffile 
pandas 
pyarrow 
numpy
tqdm
mpi4py
PyYAML
scikit-learn
torch
pytorch-lightning
tensorboard
tensorboardX